requirements.txt
agno>=2.1.4
openai>=2.3.0
python-dotenv>=1.1.1
duckduckgo-search>=6.0.0
yfinance>=0.2.40
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
lancedb>=0.10.0
pypdf>=4.0.0
fastapi>=0.110.0
uvicorn>=0.28.0
duckdb>=1.0.0
arxiv>=2.0.0
newspaper4k>=0.9.0
wikipedia>=1.4.0
google-search-results>=2.4.2
youtube-transcript-api>=0.6.0
praw>=7.7.0
streamlit>=1.30.0
arize-phoenix
openinference-instrumentation-crewai
opentelemetry-sdk
opentelemetry-exporter-otlp
